[Hardening] Minor fixes & additions
+ Copy/Move constructor for sparse array;
+ FSettings struct for App;
+ TEvent replacing Event_*Param();
+ More Copy/Move constructors;

+++ PhysicalSuika/Source/Graphics/Types.h
@@ -4,9 +4,9 @@
 
 struct FColorRGB
 {
-	int Red = 0;
-	int Green = 0;
-	int Blue = 0;
+	uint8_t Red = 0;
+	uint8_t Green = 0;
+	uint8_t Blue = 0;
 };
 
 struct FColorLinear
